Wine Quality prediction model
How can we tell if a red wine is good or not? We can use data science to find out! I used a dataset from Kaggle that has 11 features for each wine, such as acidity, alcohol, and color. The dataset also has a quality score from 0 to 10 for each wine. I made this score binary by saying that a wine is superior if it has a score of 7.56 or more, and not otherwise. Then I built different models to predict the quality of a wine based on its features. This is my project on red wine quality classification.